Added users to the system. Added tests for login/logut and story submission.
Sessions have been implemented to allow users to log in and out of the system.
Stories are now associated with users such that a user owns a given story. Also,
a user must be logged in in order to submit new stories. Added unit and
functional tests for the new system.