Implemented the front-page and the voting bin.
Redesigned the front-page of the app to show the top stories (those
with more than 5 "shoves"). Implemented the voting bin for stories
that aren't on the homepage. Added a counter cache to the association
between votes and stories. Removed the public/index.html page and
updated the routes file with our root page.