Added tests for the new user pages.
Added tests for a couple of things. The first is the counter cache that
was added to the story model (for the vote list). Secondly, added a test
to ensure the auto-voting functionality for newly submitted stories. Also,
added test for the newly remodeled index page to ensure that the correct
stories are displayed (only those that are above the threshold for votes
should appear on the homepage). In addition, tests have been added for
the voting bin page. Lastly, tests have been added to test the display
and functionality of the new user pages.